requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-google-genai
langchain-community
langchain-core
faiss-cpu
requests
yt-dlp
